Move TunerConstants into drive package
