+++ src/main/kotlin/frc/robot/Robot.kt
@@ -15,7 +15,7 @@ import edu.wpi.first.wpilibj2.command.CommandScheduler
 import frc.robot.Mode.REAL
 import frc.robot.Mode.REPLAY
 import frc.robot.Mode.SIM
-import frc.robot.generated.TunerConstants
+import frc.robot.subsystems.drive.TunerConstants
 import org.littletonrobotics.junction.LogFileUtil
 import org.littletonrobotics.junction.LoggedRobot
 import org.littletonrobotics.junction.Logger
