+++ src/main/kotlin/frc/robot/subsystems/drive/GyroIOPigeon2.java
@@ -22,7 +22,7 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.units.measure.Angle;
 import edu.wpi.first.units.measure.AngularVelocity;
-import frc.robot.generated.TunerConstants;
+
 import java.util.Queue;
 
 /** IO implementation for Pigeon 2. */
