+++ src/main/kotlin/frc/robot/subsystems/drive/ModuleIOTalonFX.java
@@ -37,7 +37,7 @@ import edu.wpi.first.units.measure.Angle;
 import edu.wpi.first.units.measure.AngularVelocity;
 import edu.wpi.first.units.measure.Current;
 import edu.wpi.first.units.measure.Voltage;
-import frc.robot.generated.TunerConstants;
+
 import java.util.Queue;
 
 /**
