+++ src/main/kotlin/frc/robot/subsystems/drive/PhoenixOdometryThread.java
@@ -18,7 +18,7 @@ import com.ctre.phoenix6.CANBus;
 import com.ctre.phoenix6.StatusSignal;
 import edu.wpi.first.units.measure.Angle;
 import edu.wpi.first.wpilibj.RobotController;
-import frc.robot.generated.TunerConstants;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Queue;
