+++ src/main/kotlin/frc/robot/subsystems/drive/TunerConstants.java
@@ -1,4 +1,4 @@
-package frc.robot.generated;
+package frc.robot.subsystems.drive;
 
 import static edu.wpi.first.units.Units.*;
 import static frc.robot.ConstantsKt.ALT_ROBORIO_SERIAL;
